+++ sprint/code/leo/precio_justo_service_factory/lib/serviceimplementation/online/online_search_strategies.rb
@@ -70,7 +70,7 @@ class OnlineFilterByPrice < OnlineFilter
 
   def initialize value
     super
-    @condition = Proc.new { |x| x.price? <= @value.to_i }
+    @condition = Proc.new { |x| x.price? <= @value.to_f }
   end
 end
